+++ megalinter/MegaLinter.py
@@ -10,7 +10,14 @@ import os
 import sys
 
 import git
-from megalinter import config, flavor_factory, linter_factory, plugin_factory, utils
+from megalinter import (
+    config,
+    flavor_factory,
+    linter_factory,
+    plugin_factory,
+    pre_post_factory,
+    utils,
+)
 from multiprocessing_logging import install_mp_handler
 
 
@@ -80,6 +87,9 @@ class Megalinter:
         self.flavor_suggestions = None
         # Initialize plugins
         plugin_factory.initialize_plugins()
+        # Run user-defined commands
+        self.pre_commands_results = pre_post_factory.run_pre_commands(self)
+        self.post_commands_results = []
         # Initialize linters and gather criteria to browse files
         self.load_linters()
         self.compute_file_extensions()
@@ -119,9 +129,17 @@ class Megalinter:
         # Update main Mega-Linter status according to results of linters run
         for linter in self.linters:
             if linter.status != "success":
-                self.status = "error"
-            if linter.return_code != 0:
+                # Not blocking linter error
+                if linter.return_code == 0:
+                    if self.status == "success":
+                        self.status = "warning"
+                # Blocking error
+                else:
+                    self.status = "error"
+            # Blocking linter error
+            if linter.return_code > 0:
                 self.return_code = linter.return_code
+            # Update number fixed
             if linter.number_fixed > 0:
                 self.has_updated_sources = 1
 
@@ -138,6 +156,9 @@ class Megalinter:
             self.flavor_suggestions = flavor_factory.get_megalinter_flavor_suggestions(
                 active_linters
             )
+
+        # Run user-defined commands
+        self.post_commands_results = pre_post_factory.run_post_commands(self)
 
         # Generate reports
         for reporter in self.reporters:
@@ -202,25 +223,28 @@ class Megalinter:
             and os.path.isdir(github_workspace + "/tmp/lint")
         ):
             logging.debug(
-                "Context: Github action run without override of DEFAULT_WORKSPACE and using /tmp/lint"
+                "[Context] Github action run without override of DEFAULT_WORKSPACE - /tmp/lint"
             )
             return github_workspace + "/tmp/lint"
         # Docker run without override of DEFAULT_WORKSPACE
         elif default_workspace != "" and os.path.isdir(
             "/tmp/lint" + os.path.sep + default_workspace
         ):
-            logging.debug("Context: Docker run without override of DEFAULT_WORKSPACE")
+            logging.debug(
+                "[Context] Docker run without override of DEFAULT_WORKSPACE"
+                f" - {default_workspace}/tmp/lint{os.path.sep + default_workspace}"
+            )
             return default_workspace + "/tmp/lint" + os.path.sep + default_workspace
         # Docker run with override of DEFAULT_WORKSPACE for test cases
         elif default_workspace != "" and os.path.isdir(default_workspace):
             logging.debug(
-                "Context: Docker run with override of DEFAULT_WORKSPACE for test cases"
+                f"[Context] Docker run test classes with override of DEFAULT_WORKSPACE - {default_workspace}"
             )
             return default_workspace
         # Docker run test classes without override of DEFAULT_WORKSPACE
         elif os.path.isdir("/tmp/lint"):
             logging.debug(
-                "Context: Docker run test classes without override of DEFAULT_WORKSPACE"
+                "[Context] Docker run test classes without override of DEFAULT_WORKSPACE - /tmp/lint"
             )
             return "/tmp/lint"
         # Github action with override of DEFAULT_WORKSPACE
@@ -229,7 +253,10 @@ class Megalinter:
             and github_workspace != ""
             and os.path.isdir(github_workspace + os.path.sep + default_workspace)
         ):
-            logging.debug("Context: Github action with override of DEFAULT_WORKSPACE")
+            logging.debug(
+                "[Context] Github action with override of DEFAULT_WORKSPACE"
+                f" - {github_workspace + os.path.sep + default_workspace}"
+            )
             return github_workspace + os.path.sep + default_workspace
         # Github action without override of DEFAULT_WORKSPACE and NOT using /tmp/lint
         elif (
@@ -239,13 +266,14 @@ class Megalinter:
             and os.path.isdir(github_workspace)
         ):
             logging.debug(
-                "Context: Github action without override of DEFAULT_WORKSPACE and NOT using /tmp/lint"
+                "[Context] Github action without override of DEFAULT_WORKSPACE and NOT using /tmp/lint"
+                f" - {github_workspace}"
             )
             return github_workspace
         # Unable to identify workspace
         else:
             raise FileNotFoundError(
-                f"Unable to find a workspace to lint \n"
+                f"[Context] Unable to find a workspace to lint \n"
                 f"DEFAULT_WORKSPACE: {default_workspace}\n"
                 f"GITHUB_WORKSPACE: {github_workspace}"
             )
@@ -360,9 +388,13 @@ class Megalinter:
         logging.debug("All found files before filtering:\n- %s", "\n- ".join(all_files))
         # Filter files according to fileExtensions, fileNames , filterRegexInclude and filterRegexExclude
         if len(self.file_extensions) > 0:
-            logging.info("- File extensions: " + ", ".join(self.file_extensions))
+            logging.info(
+                "- File extensions: " + ", ".join(sorted(self.file_extensions))
+            )
         if len(self.file_names_regex) > 0:
-            logging.info("- File names: " + ", ".join(self.file_names_regex))
+            logging.info(
+                "- File names (regex): " + ", ".join(sorted(self.file_names_regex))
+            )
         if self.filter_regex_include is not None:
             logging.info("- Including regex: " + self.filter_regex_include)
         if self.filter_regex_exclude is not None:
@@ -466,7 +498,7 @@ class Megalinter:
         logging.basicConfig(
             force=True,
             level=logging_level,
-            format="%(asctime)s [%(levelname)s] %(message)s",
+            format="%(message)s",
             handlers=[
                 logging.FileHandler(log_file, "w", "utf-8"),
                 logging.StreamHandler(sys.stdout),
@@ -507,6 +539,9 @@ class Megalinter:
         print(f"::set-output name=has_updated_sources::{str(self.has_updated_sources)}")
         if self.status == "success":
             logging.info("✅ Successfully linted all files without errors")
+            config.delete()
+        elif self.status == "warning":
+            logging.warning("◬ Successfully linted all files, but with ignored errors")
             config.delete()
         else:
             logging.error("❌ Error(s) have been found during linting")