+++ megalinter/linters/CSpellLinter.py
@@ -17,8 +17,10 @@ class CSpellLinter(Linter):
     # noinspection PyMethodMayBeStatic
     def complete_text_reporter_report(self, reporter_self):
         # Collect detected words from logs
+        if self.stdout is None:
+            return []
         whitelisted_words = []
-        for log_line in reporter_self.report_items:
+        for log_line in self.stdout.split("\n"):
             words = re.findall(r"(?<=Unknown word )\((.*)\)", log_line, re.MULTILINE)
             whitelisted_words += words
         if len(whitelisted_words) == 0:
@@ -34,6 +36,7 @@ class CSpellLinter(Linter):
                 "**/vscode-extension/**",
                 "**/.git/**",
                 ".vscode",
+                "package-lock.json",
                 "report",
             ],
             "words": whitelisted_words_clean,